% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_long.R
\name{summary_long}
\alias{summary_long}
\title{Extract basic session summary information in long format}
\usage{
summary_long(dataset = NULL)
}
\arguments{
\item{dataset}{A JSON object/list loaded into the work environment}
}
\value{
Long format table with summary session information
}
\description{
Extract basic session summary information in long format
}
\examples{
summary_long()
}
